Full python file
This python file will allow a person to type in a target webpage, search through the target sites additional webpages that use the terms "About" and "Contact" and grab both phone numbers (Australian) and email addresses. A number of improvements could be made to the project like a better regex system and a more in depth web scraper module.